MarkRec_V3.7.1-1.20
Removed Tag Along - Creating a new raijin addon pack called "Raijin Mark" or Raijin Tilt
Tweaked advs to reflect this and fixed a long standing typo oops
Removed state 0 as a result
and as a result of that combat is now much faster paced if you can become skilled with this.
crouching has now effect now.
is it better when things are simpler? sometimes perhaps.